Only try completions if installed

+++ programs/shell.nix
@@ -132,8 +132,10 @@
       rm -f ~/.zfunc/_kubectl && kubectl completion zsh > ~/.zfunc/_kubectl
       rm -f ~/.zfunc/_aws-vault &&aws-vault --completion-script-zsh  > ~/.zfunc/_aws-vault
 
-      rm -f ~/.zfunc/_rustup && rustup completions zsh > ~/.zfunc/_rustup
-      rm -f ~/.zfunc/_cargo && rustup completions zsh cargo > ~/.zfunc/_cargo
+      if [ $commands[rustup] ]; then
+        rm -f ~/.zfunc/_rustup && rustup completions zsh > ~/.zfunc/_rustup
+        rm -f ~/.zfunc/_cargo && rustup completions zsh cargo > ~/.zfunc/_cargo
+      fi
 
       complete -o nospace -C ${pkgs.terraform}/bin/terraform terraform
